+++ api/top-langs.js
@@ -1,6 +1,6 @@
 import * as dotenv from "dotenv";
 import { renderTopLanguages } from "../src/cards/top-languages-card.js";
-import blacklist from "../src/common/blacklist.js";
+import { blacklist } from "../src/common/blacklist.js";
 import {
   clampValue,
   CONSTANTS,
@@ -8,7 +8,7 @@ import {
   parseBoolean,
   renderError,
 } from "../src/common/utils.js";
-import fetchTopLanguages from "../src/fetchers/top-languages-fetcher.js";
+import { fetchTopLanguages } from "../src/fetchers/top-languages-fetcher.js";
 import { isLocaleAvailable } from "../src/translations.js";
 
 dotenv.config();
