+++ api/wakatime.js
@@ -1,5 +1,5 @@
 import * as dotenv from "dotenv";
-import wakatimeCard from "../src/cards/wakatime-card.js";
+import { renderWakatimeCard } from "../src/cards/wakatime-card.js";
 import {
   clampValue,
   CONSTANTS,
@@ -7,7 +7,7 @@ import {
   parseBoolean,
   renderError,
 } from "../src/common/utils.js";
-import fetchWakatimeStats from "../src/fetchers/wakatime-fetcher.js";
+import { fetchWakatimeStats } from "../src/fetchers/wakatime-fetcher.js";
 import { isLocaleAvailable } from "../src/translations.js";
 
 dotenv.config();
@@ -58,7 +58,7 @@ export default async (req, res) => {
     res.setHeader("Cache-Control", `public, max-age=${cacheSeconds}`);
 
     return res.send(
-      wakatimeCard(stats, {
+      renderWakatimeCard(stats, {
         custom_title,
         hide_title: parseBoolean(hide_title),
         hide_border: parseBoolean(hide_border),
