+++ jest.config.js
@@ -2,4 +2,5 @@ export default {
   clearMocks: true,
   transform: {},
   testEnvironment: "jsdom",
+  coverageProvider: "v8",
 };
