+++ package.json
@@ -24,13 +24,13 @@
     "@actions/github": "^4.0.0",
     "@testing-library/dom": "^8.17.1",
     "@testing-library/jest-dom": "^5.16.5",
-    "jest-environment-jsdom": "^29.0.3",
-    "jest": "^29.0.3",
     "@uppercod/css-to-object": "^1.1.1",
     "axios-mock-adapter": "^1.18.1",
     "color-contrast-checker": "^2.1.0",
     "hjson": "^3.2.2",
     "husky": "^4.2.5",
+    "jest": "^29.0.3",
+    "jest-environment-jsdom": "^29.0.3",
     "js-yaml": "^4.1.0",
     "lodash.snakecase": "^4.1.1",
     "parse-diff": "^0.7.0",
