+++ src/common/languageColors.json
@@ -15,6 +15,7 @@
     "ATS": "#1ac620",
     "ActionScript": "#882B0F",
     "Ada": "#02f88c",
+    "Adblock Filter List": "#800000",
     "Adobe Font Metrics": "#fa0f00",
     "Agda": "#315665",
     "Alloy": "#64C800",
@@ -22,6 +23,7 @@
     "Altium Designer": "#A89663",
     "AngelScript": "#C7D7DC",
     "Ant Build System": "#A9157E",
+    "Antlers": "#ff269e",
     "ApacheConf": "#d12127",
     "Apex": "#1797c0",
     "Apollo Guidance Computer": "#0B3D91",
@@ -41,8 +43,10 @@
     "Ballerina": "#FF5000",
     "Batchfile": "#C1F12E",
     "Beef": "#a52f4e",
+    "Berry": "#15A13C",
     "BibTeX": "#778899",
     "Bicep": "#519aba",
+    "Bikeshed": "#5562ac",
     "Bison": "#6A463F",
     "BitBake": "#00bce4",
     "Blade": "#f7523f",
@@ -52,11 +56,13 @@
     "Boo": "#d4bec1",
     "Boogie": "#c80fa0",
     "Brainfuck": "#2F2530",
+    "BrighterScript": "#66AABB",
     "Brightscript": "#662D91",
     "Browserslist": "#ffd539",
     "C": "#555555",
     "C#": "#178600",
     "C++": "#f34b7d",
+    "CAP CDS": "#0092d1",
     "CLIPS": "#00A300",
     "CMake": "#DA3434",
     "COLLADA": "#F1A42B",
@@ -66,7 +72,9 @@
     "CUE": "#5886E1",
     "CWeb": "#00007a",
     "Cabal Config": "#483465",
+    "Cadence": "#00ef8b",
     "Cairo": "#ff4a48",
+    "CameLIGO": "#3be133",
     "Cap'n Proto": "#c42727",
     "Ceylon": "#dfa535",
     "Chapel": "#8dc63f",
@@ -120,6 +128,7 @@
     "Eiffel": "#4d6977",
     "Elixir": "#6e4a7e",
     "Elm": "#60B5CC",
+    "Elvish": "#55BB55",
     "Emacs Lisp": "#c065db",
     "EmberScript": "#FFF4F3",
     "Erlang": "#B83998",
@@ -153,6 +162,9 @@
     "GSC": "#FF6800",
     "Game Maker Language": "#71b417",
     "Gemfile.lock": "#701516",
+    "Gemini": "#ff6900",
+    "Genero": "#63408e",
+    "Genero Forms": "#d8df39",
     "Genie": "#fb855d",
     "Genshi": "#951531",
     "Gentoo Ebuild": "#9400ff",
@@ -161,6 +173,7 @@
     "Gherkin": "#5B2063",
     "Git Attributes": "#F44D27",
     "Git Config": "#F44D27",
+    "Git Revision List": "#F44D27",
     "Gleam": "#ffaff3",
     "Glyph": "#c1ac7f",
     "Gnuplot": "#f0a9f0",
@@ -178,6 +191,7 @@
     "Groovy Server Pages": "#4298b8",
     "HAProxy": "#106da9",
     "HLSL": "#aace60",
+    "HOCON": "#9ff8ee",
     "HTML": "#e34c26",
     "HTML+ECR": "#2e1052",
     "HTML+EEX": "#6e4a7e",
@@ -222,6 +236,7 @@
     "JavaScript": "#f1e05a",
     "JavaScript+ERB": "#f1e05a",
     "Jest Snapshot": "#15c213",
+    "JetBrains MPS": "#21D789",
     "Jinja": "#a52a22",
     "Jison": "#56b3cb",
     "Jison Lex": "#56b3cb",
@@ -246,6 +261,7 @@
     "Latte": "#f2a542",
     "Less": "#1d365d",
     "Lex": "#DBCA00",
+    "LigoLANG": "#0e74ff",
     "LilyPond": "#9ccc7c",
     "Liquid": "#67b8de",
     "Literate Agda": "#315665",
@@ -278,13 +294,16 @@
     "Modelica": "#de1d31",
     "Modula-2": "#10253f",
     "Modula-3": "#223388",
+    "Monkey C": "#8D6747",
     "MoonScript": "#ff4585",
     "Motoko": "#fbb03b",
     "Motorola 68K Assembly": "#005daa",
+    "Move": "#4a137a",
     "Mustache": "#724b3b",
     "NCL": "#28431f",
     "NPM Config": "#cb3837",
     "NWScript": "#111522",
+    "Nasal": "#1d2c4e",
     "Nearley": "#990000",
     "Nemerle": "#3d3c6e",
     "NetLinx": "#0aa0ff",
@@ -332,6 +351,7 @@
     "PigLatin": "#fcd7de",
     "Pike": "#005390",
     "PogoScript": "#d80074",
+    "Portugol": "#f8bd00",
     "PostCSS": "#dc3a0c",
     "PostScript": "#da291c",
     "PowerBuilder": "#8f0f8d",
@@ -366,6 +386,7 @@
     "Rascal": "#fffaa0",
     "ReScript": "#ed5051",
     "Reason": "#ff5847",
+    "ReasonLIGO": "#ff5847",
     "Rebol": "#358a5b",
     "Record Jar": "#0673ba",
     "Red": "#f50000",
@@ -377,6 +398,7 @@
     "Roff": "#ecdebe",
     "Roff Manpage": "#ecdebe",
     "Rouge": "#cc0088",
+    "RouterOS Script": "#DE3941",
     "Ruby": "#701516",
     "Rust": "#dea584",
     "SAS": "#B34936",
@@ -386,6 +408,7 @@
     "SQL": "#e38c00",
     "SQLPL": "#e38c00",
     "SRecode Template": "#348a34",
+    "STL": "#373b5e",
     "SVG": "#ff9900",
     "SaltStack": "#646464",
     "Sass": "#a53b70",
@@ -425,8 +448,9 @@
     "TOML": "#9c4221",
     "TSQL": "#e38c00",
     "TSV": "#237346",
-    "TSX": "#2b7489",
+    "TSX": "#3178c6",
     "TXL": "#0178b8",
+    "Talon": "#333333",
     "Tcl": "#e4cc98",
     "TeX": "#3D6117",
     "Terra": "#00004c",
@@ -435,7 +459,7 @@
     "Thrift": "#D12127",
     "Turing": "#cf142b",
     "Twig": "#c1d026",
-    "TypeScript": "#2b7489",
+    "TypeScript": "#3178c6",
     "Unified Parallel C": "#4e3617",
     "Unity3D Asset": "#222c37",
     "Uno": "#9933cc",
@@ -446,8 +470,9 @@
     "VBScript": "#15dcdc",
     "VCL": "#148AA8",
     "VHDL": "#adb2cb",
-    "Vala": "#fbe5cd",
+    "Vala": "#a56de2",
     "Valve Data Format": "#f26025",
+    "Velocity Template Language": "#507cff",
     "Verilog": "#b2b7f8",
     "Vim Help File": "#199f4b",
     "Vim Script": "#199f4b",
@@ -458,11 +483,13 @@
     "Vyper": "#2980b9",
     "Web Ontology Language": "#5b70bd",
     "WebAssembly": "#04133b",
+    "Whiley": "#d5c397",
     "Wikitext": "#fc5757",
     "Windows Registry Entries": "#52d5ff",
     "Witcher Script": "#ff0000",
     "Wollok": "#a23738",
     "World of Warcraft Addon Data": "#f7e43f",
+    "Wren": "#383838",
     "X10": "#4B6BEF",
     "XC": "#99DA07",
     "XML": "#0060ac",
@@ -476,6 +503,7 @@
     "YARA": "#220000",
     "YASnippet": "#32AB90",
     "Yacc": "#4B6C4B",
+    "Yul": "#794932",
     "ZAP": "#0d665e",
     "ZIL": "#dc75e5",
     "ZenScript": "#00BCD1",
