+++ tests/card.test.js
@@ -1,8 +1,8 @@
 import { queryByTestId } from "@testing-library/dom";
 import "@testing-library/jest-dom";
 import { cssToObject } from "@uppercod/css-to-object";
-import Card from "../src/common/Card.js";
-import icons from "../src/common/icons.js";
+import { Card } from "../src/common/Card.js";
+import { icons } from "../src/common/icons.js";
 import { getCardColors } from "../src/common/utils.js";
 
 describe("Card", () => {
