+++ tests/fetchStats.test.js
@@ -1,8 +1,8 @@
 import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
-import calculateRank from "../src/calculateRank.js";
-import fetchStats from "../src/fetchers/stats-fetcher.js";
+import { calculateRank } from "../src/calculateRank.js";
+import { fetchStats } from "../src/fetchers/stats-fetcher.js";
 
 const data = {
   data: {
