+++ tests/pin.test.js
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import pin from "../api/pin.js";
-import renderRepoCard from "../src/cards/repo-card.js";
+import { renderRepoCard } from "../src/cards/repo-card.js";
 import { renderError } from "../src/common/utils.js";
 
 const data_repo = {
