+++ tests/retryer.test.js
@@ -1,6 +1,6 @@
 import { jest } from "@jest/globals";
 import "@testing-library/jest-dom";
-import retryer from "../src/common/retryer.js";
+import { retryer } from "../src/common/retryer.js";
 import { logger } from "../src/common/utils.js";
 
 const fetcher = jest.fn((variables, token) => {
